// generated/models/ApiMarketBatchCapabilities.ts
export class ApiMarketBatchCapabilities {
    'available': boolean;
    'execution_policy': ApiMarketBatchCapabilitiesExecutionPolicyEnum;
    'currency': ApiMarketBatchCapabilitiesCurrencyEnum;
    'payer_type': ApiMarketBatchCapabilitiesPayerTypeEnum;
    'max_orders': ApiMarketBatchCapabilitiesMaxOrdersEnum;
    'max_allocations': ApiMarketBatchCapabilitiesMaxAllocationsEnum;
    'max_calldata_bytes': ApiMarketBatchCapabilitiesMaxCalldataBytesEnum;
    'restricted_erc1155_max_order_quantity': ApiMarketBatchCapabilitiesRestrictedErc1155MaxOrderQuantityEnum;
    'requires_complete_simulation': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "available",
            "baseName": "available",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "execution_policy",
            "baseName": "execution_policy",
            "type": "ApiMarketBatchCapabilitiesExecutionPolicyEnum",
            "format": ""
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "ApiMarketBatchCapabilitiesCurrencyEnum",
            "format": ""
        },
        {
            "name": "payer_type",
            "baseName": "payer_type",
            "type": "ApiMarketBatchCapabilitiesPayerTypeEnum",
            "format": ""
        },
        {
            "name": "max_orders",
            "baseName": "max_orders",
            "type": "ApiMarketBatchCapabilitiesMaxOrdersEnum",
            "format": "int64"
        },
        {
            "name": "max_allocations",
            "baseName": "max_allocations",
            "type": "ApiMarketBatchCapabilitiesMaxAllocationsEnum",
            "format": "int64"
        },
        {
            "name": "max_calldata_bytes",
            "baseName": "max_calldata_bytes",
            "type": "ApiMarketBatchCapabilitiesMaxCalldataBytesEnum",
            "format": "int64"
        },
        {
            "name": "restricted_erc1155_max_order_quantity",
            "baseName": "restricted_erc1155_max_order_quantity",
            "type": "ApiMarketBatchCapabilitiesRestrictedErc1155MaxOrderQuantityEnum",
            "format": ""
        },
        {
            "name": "requires_complete_simulation",
            "baseName": "requires_complete_simulation",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiMarketBatchCapabilities.attributeTypeMap;
    }

    public constructor() {
    }
}

export enum ApiMarketBatchCapabilitiesExecutionPolicyEnum {
    AllOrRevert = 'ALL_OR_REVERT'
}
export enum ApiMarketBatchCapabilitiesCurrencyEnum {
    _0x0000000000000000000000000000000000000000 = '0x0000000000000000000000000000000000000000'
}
export enum ApiMarketBatchCapabilitiesPayerTypeEnum {
    Eoa = 'EOA'
}
export enum ApiMarketBatchCapabilitiesMaxOrdersEnum {
    NUMBER_128 = 128
}
export enum ApiMarketBatchCapabilitiesMaxAllocationsEnum {
    NUMBER_256 = 256
}
export enum ApiMarketBatchCapabilitiesMaxCalldataBytesEnum {
    NUMBER_1048576 = 1048576
}
export enum ApiMarketBatchCapabilitiesRestrictedErc1155MaxOrderQuantityEnum {
    _1 = '1'
}